# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Supabase Auth API",
    "version": "1.0.0",
    "description": "FlyRank A4 — Sign up, log in, log out, and protected routes using Supabase Auth (JWT)."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Credentials": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "example": "test@example.com" },
          "password": { "type": "string", "example": "password123" }
        }
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Missing email or password", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in and receive an access token (JWT)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Credentials" } } }
        },
        "responses": {
          "200": { "description": "Returns access_token and refresh_token" },
          "400": { "description": "Missing email or password", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Invalid login credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "End the user's session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out" },
          "401": { "description": "Missing or invalid token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/public/info": {
      "get": {
        "tags": ["Public"],
        "summary": "Public data — no token needed",
        "responses": { "200": { "description": "Public message" } }
      }
    },
    "/protected/profile": {
      "get": {
        "tags": ["Protected"],
        "summary": "Get the logged-in user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "User id, email, created_at" },
          "401": { "description": "Missing or invalid token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "tags": ["Protected"],
        "summary": "Dashboard for logged-in users",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Welcome message" },
          "401": { "description": "Missing or invalid token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  }
}